requirements.txt
requests~=2.23.0
Pillow~=7.1.2
PyDrive~=1.3.1
boto3~=1.12.49
python-dotenv~=0.13.0
botocore~=1.15.49
Flask~=1.1.2